+++ src/main/java/management/load/service/LocationService.java
@@ -3,9 +3,12 @@ package management.load.service;
 import lombok.AllArgsConstructor;
 import management.load.entities.Location;
 import management.load.repositories.LocationRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -17,14 +20,23 @@ public class LocationService {
     }
 
     public String deleteLocation(int id){
-        locationRepository.deleteById(id);
-        return "Delete Success";
+        Optional<Location> location = locationRepository.findById(id);
+        if(location != null) {
+            locationRepository.deleteById(id);
+            return "Delete Success";
+        }else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 
     public String updateLocation(int id,String city){
         Location location = locationRepository.findById(id).get();
-        location.setCity(city);
-        locationRepository.save(location);
-        return "Update Success";
+        if(location != null) {
+            location.setCity(city);
+            locationRepository.save(location);
+            return "Update Success";
+        }else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 }
