+++ src/main/java/management/load/service/ShipperService.java
@@ -6,9 +6,12 @@ import management.load.entities.Shipper;
 import management.load.repositories.LoadRepository;
 import management.load.repositories.ShipperRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ShipperService {
@@ -21,23 +24,43 @@ public class ShipperService {
         return shipperRepository.findAll();
     }
 
-    public List<Load> getShiperLoads(int id){
-        return loadRepository.findByShipperId(id);
+    public List<Load> getShipperLoads(int id){
+        List<Load> load = loadRepository.findByShipperId(id);
+        if(!load.isEmpty()){
+            return load;
+        }
+        else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 
     public Shipper getShipperByLocations(int locationA,int locationB){
-        return shipperRepository.getShipper(locationA,locationB);
+        Shipper shipper = shipperRepository.getShipper(locationA,locationB);
+        if(shipper != null){
+            return shipper;
+        }else{
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT);
+        }
     }
 
     public String deleteShipper(int id){
-        shipperRepository.deleteById(id);
-        return "Delete success";
+        Optional<Shipper> shipper = shipperRepository.findById(id);
+        if(shipper != null) {
+            shipperRepository.deleteById(id);
+            return "Delete success";
+        }else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 
     public String updateShipper(int id,String name){
         Shipper shipper = shipperRepository.findById(id).get();
-        shipper.setName(name);
-        shipperRepository.save(shipper);
-        return "Update success";
+        if(shipper != null) {
+            shipper.setName(name);
+            shipperRepository.save(shipper);
+            return "Update success";
+        }else{
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 }
